split cases for: - Start page - Main page
