+++ constants/base.py
@@ -1,3 +1,3 @@
 class BaseConstants:
-    DRIVER_PATH = "./drivers/chromedriver.exe"
+    DRIVER_PATH = "../drivers/chromedriver.exe"
     URL = "https://qa-complex-app-for-testing.herokuapp.com/"
